% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBiocSeed.R
\name{setBiocSeed}
\alias{setBiocSeed}
\alias{unsetBiocSeed}
\alias{enableBiocSeed}
\alias{disableBiocSeed}
\title{Smart seed setter}
\usage{
setBiocSeed(x, digits = 6)

unsetBiocSeed(info)

enableBiocSeed()

disableBiocSeed()
}
\arguments{
\item{x}{An R object of an atomic type, e.g., an integer or character vector.}

\item{digits}{Integer scalar specifying the number of significant digits to retain for double or complex \code{x}.}

\item{info}{Output of \code{setBiocSeed}.}
}
\value{
\code{setBiocSeed} will set the global seed to some deterministically chosen value based on \code{x}.
It invisibly returns the old value of \code{\link{.Random.seed}}.

\code{unsetBiocSeed} will restore the global seed to what it was before calling \code{setBiocSeed}.
It will also return \code{NULL} invisibly.

\code{disableBiocSeed} will cause all \code{setBiocSeed} and \code{unsetBiocSeed} to be no-ops.
This is reversed by \code{enableBiocSeed}.
Both functions return \code{NULL} invisibly.
}
\description{
Set the seed in a \dQuote{deterministically variable} manner,
typically for use within functions that have some stochastic component.
This reduces the pressure on the user to ensure that the seed is properly set.
}
\details{
The choice of seed is based on \code{x}, usually a chunk of the dataset of interest. 
Specifically, we convert \code{x} to an integer or character vector, and hash it to obtain an integer to use as the seed. 
This process yields a deterministic seed for reproducible results, which is also variable with different \code{x}.

Secondly, the original seed in \code{\link{.Random.seed}} is restored upon calling \code{\link{unsetBiocSeed}}.
This ensures that we do not clobber the seed in the user's session by consistently resetting it to the same value.

Thirdly, repeated invocations of \code{setBiocSeed} have no effect beyond the first.
If we have a function that calls other functions that call \code{setBiocSeed}, it can be a good idea to call \code{setBiocSeed} explicitly at the start of the outer function.
This means that the PRNG has an opportunity to progress further through the random number stream.

Finally, the entire system can be turned on or with \code{disableBiocSeed} and \code{enableBiocSeed}.
This may be useful if the seed derived from \code{x} is not appropriate and/or the user wants to control the process via the global \code{\link{set.seed}}.
}
\section{Dealing with floating-point}{

When \code{x} is double-precision, some care is required as different machines may not perform computations with the same precision.
While \code{x} might be \emph{theoretically} identical across machines, it may differ in practice in some of the less significant bits. 
These differences would result in \code{setBiocSeed} generating a machine-dependent seed, which is not desirable.

To overcome this, we round all doubles to the first \code{digits} significant figures, 
thus ignoring any machine-dependent differences in the least significant bits.
We then convert each double into two integers for the base-10 exponent and mantissa in preparation for hashing.
For complex numbers, the same treatment is applied to the real and imaginary parts separately.
}

\examples{
info <- setBiocSeed(letters)
runif(5)
unsetBiocSeed(info)

# Same as above.
info <- setBiocSeed(letters)
runif(5)
unsetBiocSeed(info)

# Different, but deterministically so.
info <- setBiocSeed(LETTERS)
runif(5)
unsetBiocSeed(info)

# Does not affect the RNG outside of the set/unset pair.
info <- setBiocSeed(LETTERS)
X <- runif(5)
unsetBiocSeed(info)
runif(1)

info <- setBiocSeed(LETTERS)
X <- runif(5)
unsetBiocSeed(info)
runif(1)

# Typical usage in a function:
FUN <- function(x) {
    info <- setBiocSeed(head(x))
    on.exit(unsetBiocSeed(info))
    sample(x)
}

FUN(1:10)
FUN(1:10)

}
\author{
Aaron Lun
}
